+++ codes/misc/bit-hacks.hpp
@@ -1,9 +1,9 @@
 // next permutation of x as a bit sequence
-ull next_bits_permutation(ull x) {
-  ull c = __builtin_ctzll(x), r = x + (1ULL << c);
+u64 next_bits_permutation(u64 x) {
+  u64 c = __builtin_ctzll(x), r = x + (u64(1) << c);
   return (r ^ x) >> (c + 2) | r;
 }
 // iterate over all (proper) subsets of bitset s
-void subsets(ull s) {
-  for (ull x = s; x;) { --x &= s; /* do stuff */ }
+void subsets(u64 s) {
+  for (u64 x = s; x;) { --x &= s; /* do stuff */ }
 }
